+++ src/file-io.c
@@ -1,6 +1,6 @@
 #include <petsc.h>
 #include <petscviewerhdf5.h>
-#include "hybrid.h"
+#include "ecsprt.h"
 
 
 PetscErrorCode LoadFluidQuantities(PetscReal fluxScale[NDIM], char inpath[PETSC_MAX_PATH_LEN], Context *ctx)