+++ src/solver.h
@@ -1,6 +1,6 @@
 #ifndef SOLVER_H
 #define SOLVER_H
 
-#include "hybrid.h"
+#include "ecsprt.h"
 
 #endif // SOLVER_H